scripts/2_scrape_pages.py
#!/usr/bin/env python3
"""
TERVYX Amazon Claims — Page Scraping (Ethics-First)
Fetch public product pages with robots.txt compliance and Wayback archival.
"""

import argparse
import csv
import hashlib
import time
import yaml
import requests
from pathlib import Path
from typing import Dict, Optional
from datetime import datetime
from urllib.parse import urljoin
from bs4 import BeautifulSoup


def load_scraping_policy(policy_path: Path) -> dict:
    """Load scraping policy YAML."""
    with open(policy_path) as f:
        return yaml.safe_load(f)


def check_robots_txt(domain: str, user_agent: str) -> bool:
    """
    Check robots.txt compliance.

    TODO: Implement proper robots.txt parsing
    - Use robotparser module
    - Cache results per domain
    - Respect Crawl-delay directive
    """
    # FIXME: Replace with actual check
    print(f"  [robots.txt] Checking {domain}... (assuming allowed)")
    return True


def save_to_wayback(url: str, policy: dict) -> Optional[str]:
    """
    Save URL to Wayback Machine.

    TODO: Implement Wayback API call
    - POST to https://web.archive.org/save/{url}
    - Handle rate limits
    - Return archive URL
    """
    if not policy.get("wayback_save", False):
        return None

    # FIXME: Replace with actual API call
    wayback_url = f"https://web.archive.org/web/20251112000000/{url}"
    print(f"  [wayback] Archived (simulated): {wayback_url}")
    return wayback_url


def fetch_page(url: str, policy: dict) -> Dict:
    """
    Fetch product page with throttling and compliance.

    TODO: Implement robust fetching
    - User-Agent header
    - Timeout, retries with exponential backoff
    - Error handling (404, 403, etc.)
    - Respect rate limits
    """
    headers = {"User-Agent": policy["user_agent"]}
    timeout = policy["timeout_seconds"]

    # FIXME: Replace with actual fetch
    print(f"  [fetch] {url}")
    time.sleep(policy["throttle_seconds"])  # Respect throttle

    # Simulated response
    html = f"<html><body><h1>Product Page</h1><p>Claim: Improves sleep quality by 87%</p></body></html>"
    sha256 = hashlib.sha256(html.encode()).hexdigest()

    return {
        "html": html,
        "sha256": sha256,
        "status_code": 200,
        "captured_at": datetime.utcnow().isoformat() + "Z"
    }


def extract_assets(html: str, base_url: str) -> list:
    """
    Extract image URLs from HTML.

    TODO: Parse product images
    - Find main image + gallery
    - Filter for claim-containing images
    - Return list of asset URLs
    """
    # FIXME: Replace with actual parsing
    soup = BeautifulSoup(html, "lxml")
    images = []

    for img in soup.find_all("img")[:3]:  # Limit for demo
        src = img.get("src", "")
        if src:
            images.append({
                "url": urljoin(base_url, src),
                "alt": img.get("alt", "")
            })

    return images


def main():
    parser = argparse.ArgumentParser(description="Scrape product pages")
    parser.add_argument("--in", dest="input_csv", type=Path, required=True)
    parser.add_argument("--out", type=Path, required=True, help="SHA256 index CSV")
    parser.add_argument("--assets", type=Path, required=True, help="Assets index CSV")
    parser.add_argument("--policy", type=Path, required=True)
    args = parser.parse_args()

    policy = load_scraping_policy(args.policy)

    print(f"Scraping from {args.input_csv}")
    print(f"Policy: throttle={policy['throttle_seconds']}s, max_retries={policy['max_retries']}")

    pages = []
    assets = []

    with open(args.input_csv) as f:
        reader = csv.DictReader(f)
        for i, row in enumerate(reader):
            asin = row["asin"]
            url = row["url"]

            print(f"\n[{i+1}] {asin}: {url}")

            # Check robots.txt
            if not check_robots_txt("www.amazon.com", policy["user_agent"]):
                print("  [SKIP] robots.txt disallows")
                continue

            # Fetch page
            result = fetch_page(url, policy)

            # Wayback archive
            wayback_url = save_to_wayback(url, policy)

            pages.append({
                "asin": asin,
                "page_sha256": result["sha256"],
                "wayback_url": wayback_url or "",
                "captured_at": result["captured_at"],
                "status_code": result["status_code"]
            })

            # Extract assets
            page_assets = extract_assets(result["html"], url)
            for j, asset in enumerate(page_assets):
                assets.append({
                    "asin": asin,
                    "asset_id": f"{asin}_img{j:02d}",
                    "asset_type": "image",
                    "url": asset["url"],
                    "wayback_url": "",  # TODO: Archive images
                    "sha256": hashlib.sha256(asset["url"].encode()).hexdigest(),  # Full 64-char hash
                    "storage_uri": "",
                    "width": "",
                    "height": ""
                })

    # Write outputs
    args.out.parent.mkdir(parents=True, exist_ok=True)
    with open(args.out, "w", newline="") as f:
        fieldnames = ["asin", "page_sha256", "wayback_url", "captured_at", "status_code"]
        writer = csv.DictWriter(f, fieldnames=fieldnames)
        writer.writeheader()
        writer.writerows(pages)

    args.assets.parent.mkdir(parents=True, exist_ok=True)
    with open(args.assets, "w", newline="") as f:
        fieldnames = ["asin", "asset_id", "asset_type", "url", "wayback_url", "sha256", "storage_uri", "width", "height"]
        writer = csv.DictWriter(f, fieldnames=fieldnames)
        writer.writeheader()
        writer.writerows(assets)

    print(f"\n✓ Scraped {len(pages)} pages")
    print(f"✓ Indexed {len(assets)} assets")
    print(f"✓ Saved to {args.out}, {args.assets}")


if __name__ == "__main__":
    main()
